requirements.txt
anthropic>=0.52.0
openai>=1.70.0
aiohttp>=3.11.0
python-dotenv>=1.0.0
fastmcp>=3.0.2
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
aioresponses>=0.7.6